Plans 2023: Compile manageHref out of the grid components
